Chore: Refactor database.
