+++ tests/lib/microservices/sync-service/sync-service.ts
@@ -14,9 +14,11 @@ const Queue = function () {
 const queueObject = { Queue };
 const database = {
     connect() { },
-    getJob() { },
-    unlock() { },
-    updateJob() { }
+    job: {
+        get() { },
+        update() { }
+    },
+    unlock() { }
 };
 
 const data = {
@@ -43,7 +45,7 @@ test.beforeEach(async (t) => {
     sinon.stub(resultsQueue, 'listen').resolves();
     sinon.stub(database, 'lock').resolves();
     sinon.stub(database, 'unlock').resolves();
-    sinon.stub(database, 'updateJob').resolves();
+    sinon.stub(database.job, 'update').resolves();
 
     t.context.job = JSON.parse(await readFileAsync(path.join(__dirname, 'fixtures', 'dbdata.json')));
 
@@ -59,12 +61,12 @@ test.afterEach.always((t) => {
     t.context.resultsQueue.listen.restore();
     t.context.database.lock.restore();
     t.context.database.unlock.restore();
-    t.context.database.updateJob.restore();
-    t.context.database.getJob.restore();
+    t.context.database.job.update.restore();
+    t.context.database.job.get.restore();
 });
 
 test.serial(`if a job doesn't exists in database, it should report an error and unlock the key`, async (t) => {
-    sinon.stub(database, 'getJob').resolves();
+    sinon.stub(database.job, 'get').resolves();
     sinon.spy(logger, 'error');
 
     await sync.run();
@@ -74,14 +76,14 @@ test.serial(`if a job doesn't exists in database, it should report an error and 
     t.true(t.context.logger.error.calledOnce);
     t.true(t.context.database.lock.calledOnce);
     t.true(t.context.database.unlock.calledOnce);
-    t.false(t.context.database.updateJob.called);
+    t.false(t.context.database.job.update.called);
 
     t.context.logger.error.restore();
 });
 
 test.serial(`if the job in the database has the status 'error', it should work as normal`, async (t) => {
     t.context.job.status = JobStatus.error;
-    sinon.stub(database, 'getJob').resolves(t.context.job);
+    sinon.stub(database.job, 'get').resolves(t.context.job);
 
     await sync.run();
 
@@ -89,11 +91,11 @@ test.serial(`if the job in the database has the status 'error', it should work a
 
     t.true(t.context.database.lock.calledOnce);
     t.true(t.context.database.unlock.calledOnce);
-    t.true(t.context.database.updateJob.called);
+    t.true(t.context.database.job.update.called);
 });
 
 test.serial(`if the job status is 'started' and the job status is database 'pending', it should update the status and the started property`, async (t) => {
-    sinon.stub(database, 'getJob').resolves(t.context.job);
+    sinon.stub(database.job, 'get').resolves(t.context.job);
 
     await sync.run();
 
@@ -101,8 +103,8 @@ test.serial(`if the job status is 'started' and the job status is database 'pend
 
     t.true(t.context.database.lock.calledOnce);
     t.true(t.context.database.unlock.calledOnce);
-    t.true(t.context.database.updateJob.called);
-    const dbJob: IJob = t.context.database.updateJob.args[0][0];
+    t.true(t.context.database.job.update.called);
+    const dbJob: IJob = t.context.database.job.update.args[0][0];
 
     t.is(dbJob.status, JobStatus.started);
     t.is(dbJob.started, data.started.started);
@@ -110,7 +112,7 @@ test.serial(`if the job status is 'started' and the job status is database 'pend
 
 test.serial(`if the job status is 'started' and the job status in database is not 'pending', it should update just the started property`, async (t) => {
     t.context.job.status = JobStatus.finished;
-    sinon.stub(database, 'getJob').resolves(t.context.job);
+    sinon.stub(database.job, 'get').resolves(t.context.job);
 
     await sync.run();
 
@@ -118,8 +120,8 @@ test.serial(`if the job status is 'started' and the job status in database is no
 
     t.true(t.context.database.lock.calledOnce);
     t.true(t.context.database.unlock.calledOnce);
-    t.true(t.context.database.updateJob.called);
-    const dbJob: IJob = t.context.database.updateJob.args[0][0];
+    t.true(t.context.database.job.update.called);
+    const dbJob: IJob = t.context.database.job.update.args[0][0];
 
     t.is(dbJob.status, JobStatus.finished);
     t.is(dbJob.started, data.started.started);
@@ -128,7 +130,7 @@ test.serial(`if the job status is 'started' and the job status in database is no
 test.serial(`if the job status is 'started' and the property started in database is greater than the current one, it should update the started property`, async (t) => {
     t.context.job.status = JobStatus.finished;
     t.context.job.started = new Date('2017-08-31T23:55:00.877Z');
-    sinon.stub(database, 'getJob').resolves(t.context.job);
+    sinon.stub(database.job, 'get').resolves(t.context.job);
 
     await sync.run();
 
@@ -136,15 +138,15 @@ test.serial(`if the job status is 'started' and the property started in database
 
     t.true(t.context.database.lock.calledOnce);
     t.true(t.context.database.unlock.calledOnce);
-    t.true(t.context.database.updateJob.called);
-    const dbJob: IJob = t.context.database.updateJob.args[0][0];
+    t.true(t.context.database.job.update.called);
+    const dbJob: IJob = t.context.database.job.update.args[0][0];
 
     t.is(dbJob.status, JobStatus.finished);
     t.is(dbJob.started, data.started.started);
 });
 
 test.serial(`if the job status is 'error', it should update the job in database properly`, async (t) => {
-    sinon.stub(database, 'getJob').resolves(t.context.job);
+    sinon.stub(database.job, 'get').resolves(t.context.job);
 
     await sync.run();
 
@@ -152,8 +154,8 @@ test.serial(`if the job status is 'error', it should update the job in database 
 
     t.true(t.context.database.lock.calledOnce);
     t.true(t.context.database.unlock.calledOnce);
-    t.true(t.context.database.updateJob.called);
-    const dbJob: IJob = t.context.database.updateJob.args[0][0];
+    t.true(t.context.database.job.update.called);
+    const dbJob: IJob = t.context.database.job.update.args[0][0];
 
     t.not(dbJob.status, JobStatus.error);
     t.is(dbJob.finished, data.error.finished);
@@ -161,7 +163,7 @@ test.serial(`if the job status is 'error', it should update the job in database 
 });
 
 test.serial(`if the job status is 'finished' and all rules are processed, it should update rules and send the status finished if there is no errors`, async (t) => {
-    sinon.stub(database, 'getJob').resolves(t.context.job);
+    sinon.stub(database.job, 'get').resolves(t.context.job);
 
     await sync.run();
 
@@ -169,9 +171,9 @@ test.serial(`if the job status is 'finished' and all rules are processed, it sho
 
     t.true(t.context.database.lock.calledOnce);
     t.true(t.context.database.unlock.calledOnce);
-    t.true(t.context.database.updateJob.called);
+    t.true(t.context.database.job.update.called);
 
-    const dbJob: IJob = t.context.database.updateJob.args[0][0];
+    const dbJob: IJob = t.context.database.job.update.args[0][0];
 
     t.is(dbJob.status, JobStatus.finished);
     t.is(dbJob.finished, data.finished.finished);
@@ -183,7 +185,7 @@ test.serial(`if the job status is 'finished' and all rules are processed, it sho
 
 test.serial(`if the job status is 'finished' and all rules are processed, it should update rules and send the status error if there is a previous error in database`, async (t) => {
     t.context.job.error = [data.error.error];
-    sinon.stub(database, 'getJob').resolves(t.context.job);
+    sinon.stub(database.job, 'get').resolves(t.context.job);
 
     await sync.run();
 
@@ -191,9 +193,9 @@ test.serial(`if the job status is 'finished' and all rules are processed, it sho
 
     t.true(t.context.database.lock.calledOnce);
     t.true(t.context.database.unlock.calledOnce);
-    t.true(t.context.database.updateJob.called);
+    t.true(t.context.database.job.update.called);
 
-    const dbJob: IJob = t.context.database.updateJob.args[0][0];
+    const dbJob: IJob = t.context.database.job.update.args[0][0];
 
     t.is(dbJob.status, JobStatus.error);
     t.is(dbJob.finished, data.finished.finished);
@@ -204,7 +206,7 @@ test.serial(`if the job status is 'finished' and all rules are processed, it sho
 });
 
 test.serial(`if the job status is 'finished' and all rules are processed, it should update rules and send the status error if there is any error`, async (t) => {
-    sinon.stub(database, 'getJob').resolves(t.context.job);
+    sinon.stub(database.job, 'get').resolves(t.context.job);
 
     await sync.run();
 
@@ -212,9 +214,9 @@ test.serial(`if the job status is 'finished' and all rules are processed, it sho
 
     t.true(t.context.database.lock.calledOnce);
     t.true(t.context.database.unlock.calledOnce);
-    t.true(t.context.database.updateJob.called);
+    t.true(t.context.database.job.update.called);
 
-    const dbJob: IJob = t.context.database.updateJob.args[0][0];
+    const dbJob: IJob = t.context.database.job.update.args[0][0];
 
     t.is(dbJob.status, JobStatus.error);
     t.is(dbJob.finished, data.finished.finished);
@@ -225,37 +227,37 @@ test.serial(`if the job status is 'finished' and all rules are processed, it sho
 });
 
 test.serial(`if the job status is 'finished' but they are partial results, it should update rules and just send the status finished when all the rules are processed`, async (t) => {
-    sinon.stub(database, 'getJob').resolves(t.context.job);
+    sinon.stub(database.job, 'get').resolves(t.context.job);
 
     await sync.run();
 
     await t.context.resultsQueue.listen.args[0][0]([data.started]);
 
-    let dbJob: IJob = t.context.database.updateJob.args[0][0];
+    let dbJob: IJob = t.context.database.job.update.args[0][0];
 
     t.is(dbJob.status, JobStatus.started);
     t.is(dbJob.started, data.started.started);
 
     await t.context.resultsQueue.listen.args[0][0]([data.finishedPart1]);
 
-    dbJob = t.context.database.updateJob.args[1][0];
+    dbJob = t.context.database.job.update.args[1][0];
 
     t.is(dbJob.status, JobStatus.started);
 
     await t.context.resultsQueue.listen.args[0][0]([data.finishedPart2]);
 
-    dbJob = t.context.database.updateJob.args[2][0];
+    dbJob = t.context.database.job.update.args[2][0];
 
     t.is(dbJob.status, JobStatus.finished);
     t.truthy(dbJob.finished);
 
     t.is(t.context.database.lock.callCount, 3);
     t.is(t.context.database.unlock.callCount, 3);
-    t.is(t.context.database.updateJob.callCount, 3);
+    t.is(t.context.database.job.update.callCount, 3);
 });
 
 test.serial(`if the job receive more than one message from the same id, it should lock the database just once`, async (t) => {
-    sinon.stub(database, 'getJob').resolves(t.context.job);
+    sinon.stub(database.job, 'get').resolves(t.context.job);
 
     await sync.run();
 
@@ -263,11 +265,11 @@ test.serial(`if the job receive more than one message from the same id, it shoul
 
     t.is(t.context.database.lock.callCount, 1);
     t.is(t.context.database.unlock.callCount, 1);
-    t.is(t.context.database.updateJob.callCount, 1);
+    t.is(t.context.database.job.update.callCount, 1);
 });
 
 test.serial(`if the job receive two messages with different id, it should lock the database twice`, async (t) => {
-    sinon.stub(database, 'getJob').resolves(t.context.job);
+    sinon.stub(database.job, 'get').resolves(t.context.job);
 
     await sync.run();
 
@@ -275,5 +277,5 @@ test.serial(`if the job receive two messages with different id, it should lock t
 
     t.is(t.context.database.lock.callCount, 2);
     t.is(t.context.database.unlock.callCount, 2);
-    t.is(t.context.database.updateJob.callCount, 2);
+    t.is(t.context.database.job.update.callCount, 2);
 });
